Added file parsing for *.apj and Physical.pkg
Now the AS version used in the project is not hard coded anymore, but
parsed from the project file.
Also configurations are not anymore all directories within the Physical
folder, but parsed from the Physical.pkg

Additionally the descriptions of the projects and the configurations are
also parsed from the files and presented in the selection dialogs.