Added notification after extension update, fixes #26
A new module for notifications was introduced which handles all pop-ups. To know which version was already shown, the vscode globalState was used and is synchronized with settings sync.
A module to handle the global state was added which works with a singleton object providing getters and setters for all global and workspace state.
The BrConfiguration.ts module was also partially adjusted to use the new pattern with getters and setters.