Split up build options from AS configuration files
Before, the build options from the Cpu.pkg files were handed over to the C/C++ extension as a single string.
The C/C++ extension handled it as a single argument and escaped it with "" in the shell call.
When using multiple build options this lead to compiler errors which prevented IntelliSense to work as expected.

With this commit the build options are split on each whitespace and handed to the C/C++ extension as a string array.
Now multiple options can be used as expected and -W or -profile options do not lead to an error.

Fixes #27