Improved Automation Studio version information parsing
The ProductInfo.ini file which was used to parse the AS version does only exist after a service pack was installed.
In newer AS versions >= 4.11 the file does not exist anymore, even after an SP installation.
Therefore the ProductInfo_bin-xx.brv file is now used as a source for version information.